+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [1.0.0] - 2019-02-27
+
+### Added
+- [#3](https://github.com/cloudcreativity/json-api-testing/issues/3)
+Can now assert exact errors on a document and HTTP message.
+- [#5](https://github.com/cloudcreativity/json-api-testing/issues/5)
+Can now assert a server generated id with a known id.
+
+### Fixed
+- [#6](https://github.com/cloudcreativity/json-api-testing/issues/6)
+Fixed asserting that an error exists with only an integer HTTP status code.
+- [#4](https://github.com/cloudcreativity/json-api-testing/issues/4)
+Fixed incorrect diff when asserting a resource identifier on a document that contains a resource object.
+
 ## [1.0.0-rc.1] - 2019-01-03
 
 ### Added